fedimint_mint_server
Module db

Structs

DbKeyPrefixIter

An iterator over the variants of DbKeyPrefix

ECashUserBackupSnapshot

User’s backup, received at certain time, containing encrypted payload

EcashBackupKey

Key used to store user’s ecash backups

EcashBackupKeyPrefix

MintAuditItemKeyPrefix

MintOutputOutcomeKey

Transaction id and output index identifying an output outcome

MintOutputOutcomePrefix

NonceKey

NonceKeyPrefix

Enums

DbKeyPrefix

MintAuditItemKey

Represents the amounts of issued (signed) and redeemed (verified) notes for auditing