Struct fedimint_mint_server::Mint

pub struct Mint {
    pub(crate) cfg: MintConfig,
    pub(crate) sec_key: Tiered<SecretKeyShare>,
    pub(crate) pub_key_shares: BTreeMap<PeerId, Tiered<PublicKeyShare>>,
    pub(crate) pub_key: HashMap<Amount, AggregatePublicKey>,
}

Federated mint member mint

Fields
cfg: MintConfigsec_key: Tiered<SecretKeyShare>pub_key_shares: BTreeMap<PeerId, Tiered<PublicKeyShare>>pub_key: HashMap<Amount, AggregatePublicKey>
Implementations

impl Mint

pub(crate) async fn handle_backup_request(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>,
request: SignedBackupRequest
) -> Result<(), ApiError>

pub(crate) async fn handle_recover_request(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>,
id: XOnlyPublicKey
) -> Option<ECashUserBackupSnapshot>

impl Mint

pub fn new(cfg: MintConfig) -> Mint

Constructs a new mint

Panics

If there are no amount tiers

If the amount tiers for secret and public keys are inconsistent

If the pub key belonging to the secret key share is not in the pub key list.

pub fn pub_key(&self) -> HashMap<Amount, AggregatePublicKey>

pub(crate) fn blind_sign(
&self,
output: TieredMulti<BlindNonce>
) -> Result<MintOutputSignatureShare, MintError>